@@ -38,20 +38,22 @@
     "babel-eslint": "^4.1.3",
     "babel-jest": "^5.3.0",
     "babel-loader": "^5.3.2",
-    "babel-plugin-react-transform": "^1.0.5",
-    "eslint": "^1.5.0",
-    "eslint-config-airbnb": "0.0.8",
-    "eslint-plugin-react": "^3.4.2",
-    "jest-cli": "^0.5.7",
-    "react": "^0.13.3",
+    "babel-plugin-react-transform": "^1.1.1",
+    "eslint": "^1.6.0",
+    "eslint-config-airbnb": "0.1.0",
+    "eslint-plugin-react": "^3.5.1",
+    "jest-cli": "^0.5.10",
+    "react": "^0.14.0",
+    "react-addons-test-utils": "^0.14.0",
+    "react-dom": "^0.14.0",
     "react-transform-catch-errors": "^1.0.0",
-    "react-transform-hmr": "^1.0.0",
-    "redbox-react": "^1.0.6",
+    "react-transform-hmr": "^1.0.1",
+    "redbox-react": "^1.1.1",
     "underscore": "^1.8.3",
     "webpack": "^1.12.2",
     "webpack-dev-middleware": "^1.2.0",
-    "webpack-dev-server": "^1.11.0",
-    "webpack-hot-middleware": "^2.2.0"
+    "webpack-dev-server": "^1.12.0",
+    "webpack-hot-middleware": "^2.4.1"
   },
   "jest": {
     "unmockedModulePathPatterns": [
