@@ -1,16 +1,16 @@
 import React from 'react';
+import ReactDOM from 'react-dom';
 import routes from './components/routes';
 import Router from 'react-router';
 import initializeTheme from './theme/initialize';
-
-import belle from 'belle';
-
+import { createHashHistory } from 'history';
 
 // export for http://fb.me/react-devtools
 window.React = React;
 
 initializeTheme();
 
-Router.run(routes, (Handler) => {
-  React.render(<Handler/>, document.getElementById('react'));
-});
+const history = createHashHistory();
+const rootComponent = <Router history={history} children={routes} />;
+
+ReactDOM.render(rootComponent, document.getElementById('react'));
