@@ -11,8 +11,8 @@ app.use('/icon', express.static('icon'));
 app.use('/images', express.static('images'));
 
 app.use(require('webpack-dev-middleware')(compiler, {
-    noInfo: true,
-    publicPath: config.output.publicPath
+  noInfo: true,
+  publicPath: config.output.publicPath
 }));
 
 app.use(require('webpack-hot-middleware')(compiler));
@@ -21,7 +21,7 @@ app.get('*', function(req, res) {
   res.sendFile(path.join(__dirname, 'index.html'));
 });
 
-app.listen(3000, 'localhost', function (err) {
+app.listen(3000, 'localhost', function(err) {
   if (err) {
     console.log(err);
     return;
