@@ -3,8 +3,8 @@
 jest.dontMock('../components/Card');
 jest.dontMock('../utils/inject-style');
 
-import React from 'react/addons';
-const TestUtils = React.addons.TestUtils;
+import React from 'react';
+import TestUtils from 'react-addons-test-utils';
 
 // Babel would move an import in front of the jest.dontMock. That's why require
 // is used instead of import.
@@ -13,12 +13,11 @@ const Card = require('../components/Card');
 describe('Card', () => {
   it('should come with default styles', () => {
     const card = TestUtils.renderIntoDocument(
-      <Card/>
+      <Card />
     );
     const divNode = TestUtils.findRenderedDOMComponentWithTag(card, 'div');
-
-    expect(divNode.props.style).toBeDefined();
-    expect(divNode.props.style.background).toBe('#fff');
+    expect(divNode.hasAttribute('style')).toBeTruthy();
+    expect(divNode.getAttribute('style').indexOf('background:#fff') > -1).toBeTruthy();
   });
 
   it('should be able to adopt the style of the card', () => {
@@ -27,7 +26,7 @@ describe('Card', () => {
     );
     const divNode = TestUtils.findRenderedDOMComponentWithTag(card, 'div');
 
-    expect(divNode.props.style.background).toEqual('#F00');
+    expect(divNode.getAttribute('style').indexOf('background:#F00') > -1).toBeTruthy();
   });
 
   it('should render its children', () => {
@@ -37,7 +36,8 @@ describe('Card', () => {
     const divNode = TestUtils.findRenderedDOMComponentWithTag(card, 'div');
     const spanNode = TestUtils.findRenderedDOMComponentWithTag(card, 'span');
 
-    expect(divNode.props.children.type).toEqual('span');
-    expect(spanNode.props.children).toEqual('Hello there');
+    expect(divNode.tagName).toEqual('DIV');
+    expect(divNode.childNodes[0].tagName).toEqual('SPAN');
+    expect(spanNode.textContent).toEqual('Hello there');
   });
 });
