@@ -4,8 +4,8 @@ jest.dontMock('../components/ComboBox');
 jest.dontMock('../components/Option');
 
 import {extend} from 'underscore';
-import React from 'react/addons';
-const TestUtils = React.addons.TestUtils;
+import React from 'react';
+import TestUtils from 'react-addons-test-utils';
 
 // Babel would move an import in front of the jest.dontMock. That's why require
 // is used instead of import.
@@ -125,7 +125,8 @@ describe('ComboBox', () => {
     );
 
     const selectedAreaNode = TestUtils.scryRenderedDOMComponentsWithTag(combobox, 'input')[1];
-    expect(selectedAreaNode.props.style.cursor).toBe('cross');
+    expect(selectedAreaNode.hasAttribute('style')).toBeTruthy();
+    expect(selectedAreaNode.getAttribute('style').indexOf('cursor:cross') > -1).toBeTruthy();
   });
 
   describe('updating props', () => {
