@@ -1,16 +1,14 @@
 /**
- * Returns true if the provided property is a component of the provided name.
+ * Returns true if the provided element is a component of the provided type.
  *
- * @param componentName {String} - the component's name of the class definition
+ * @param classType {ReactElement class} - the class of a React Element
  * @param reactElement {ReactElement} - any React Element (not a real DOM node)
  *
  * @example
  * // Checks if the component is an Autocomplete
- * isComponentType('Autocomplete', this.props.children[0]);
+ * isComponentType(Autocomplete, this.props.children[0]);
  */
-export default function isComponentOfType(displayName, reactElement) {
+export default function isComponentOfType(classType, reactElement) {
   return reactElement &&
-         reactElement._isReactElement &&
-         reactElement.type &&
-         reactElement.type.displayName === displayName;
+         reactElement.type === classType;
 }
